Convert to ES6 classes
