+++ src/BreakError.js
@@ -1,0 +1,4 @@
+class BreakError extends Error {
+}
+
+module.exports = BreakError;
